Move osc_server class to own file
